# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: 'Agrocom API'
  description: 'API del Sistema de Gestión de Operaciones de Fumigación de Agrocom SRL. Sirve a las apps de campo (`agrocom-field`) y al portal del cliente; el spec generado desde el código es el contrato de la API (ADR 0014). Dinero y hectáreas viajan como string decimal, nunca como número de punto flotante.'
  version: 0.1.0
servers:
  -
    url: /
    description: 'Servidor actual — la URL base la define el entorno (local, staging, producción).'
paths:
...
  /api/evidencias:
    post:
      tags:
        - Operaciones
      summary: 'Sube una evidencia (multipart/form-data)'
      description: 'Sube un archivo de evidencia (captura de RC, foto de campo, foto de incidencia, comprobante o firma de acta) y responde `aplicado`, `duplicado` (reintento del mismo `uuid_cliente`, se trata como éxito) o `rechazado` con motivo — mismo vocabulario que `POST /api/sync`, aunque este endpoint vive aparte porque su "sobre" lleva un binario. El servidor SIEMPRE recalcula el hash SHA-256 sobre el contenido recibido (es el que queda persistido); si el dispositivo declara `hash_dispositivo` (calculado al capturar, ADR 0009), se compara contra ese cálculo y se rechaza si no coincide — detecta corrupción en tránsito, sin confiar ciegamente en el valor declarado. La compresión a <300 KB y la cola de reintentos son responsabilidad de la app de campo (TE-07 lado app, repo `agrocom-field`) — este endpoint no las repite.'
      operationId: subirEvidencia
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              required:
                - uuid_cliente
                - tipo
                - fecha
                - archivo
              properties:
                uuid_cliente:
                  type: string
                  example: a1b2c3d4-0000-4000-8000-000000000010
                tipo:
                  type: string
                  example: imagen_campo
                  enum:
                    - captura_rc
                    - imagen_campo
                    - foto_incidencia
                    - comprobante
                    - firma_acta
                    - foto_control
                    - foto_ciclo_bateria_balanceo
                    - foto_dron_limpio
                fecha:
                  type: string
                  format: date-time
                  example: '2026-09-01T10:00:00-04:00'
                hash_dispositivo:
                  description: 'SHA-256 calculado en el dispositivo al capturar (ADR 0009), opcional — si se declara, debe coincidir con el hash recalculado por el servidor.'
                  type: string
                  example: null
                  nullable: true
                archivo:
                  type: string
                  format: binary
              type: object
      responses:
        200:
          description: 'Resultado de aplicar el registro.'
          content:
            application/json:
              schema:
                required:
                  - uuid_cliente
                  - estado
                properties:
                  uuid_cliente:
                    type: string
                    example: a1b2c3d4-0000-4000-8000-000000000010
                    nullable: true
                  estado:
                    type: string
                    example: aplicado
                    enum:
                      - aplicado
                      - duplicado
                      - rechazado
                  motivo:
                    description: 'Presente solo cuando `estado` es `rechazado`.'
                    type: string
                    example: 'archivo ausente o vacío'
                type: object
        401:
          description: 'Token ausente, revocado o ya sin rol válido.'
        422:
          description: 'El archivo supera el tope de tamaño de plataforma (20 MB).'
      security:
        -
          tokenDispositivo: []
...
  /api/sync:
    post:
      tags:
        - Sincronizacion
      summary: 'Push de sincronización en lote (trabajo, sesión)'
      description: 'Push en lote de la cola offline (espec §2.1). Cada registro se procesa en su propia transacción —nunca el lote completo en una— y se responde con uno de tres estados: `aplicado`, `duplicado` (reintento del mismo `uuid_cliente` ya aplicado, se trata como éxito) o `rechazado` con motivo. El servidor agrupa por tipo y aplica siempre en el orden `trabajo`, `sesion`, `cierre_trabajo`, `cierre_sesion`, sin importar el orden del arreglo recibido, así que un registro puede referenciar otro del mismo lote aunque venga antes en el arreglo.'
      operationId: sincronizarLote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - registros
              properties:
                registros:
                  type: array
                  items:
                    $ref: '#/components/schemas/RegistroSync'
              type: object
      responses:
        200:
          description: 'Resultado por registro, en el mismo orden del arreglo de entrada.'
          content:
            application/json:
              schema:
                required:
                  - resultados
                properties:
                  resultados:
                    type: array
                    items:
                      $ref: '#/components/schemas/ResultadoSync'
                type: object
        401:
          description: 'Token ausente, revocado o ya sin rol válido.'
        422:
          description: 'Falta `registros`, o no es un arreglo.'
      security:
        -
          tokenDispositivo: []
components:
  schemas:
...
    RegistroSync:
      title: 'Registro de entrada del lote de sync'
      description: 'Un elemento del arreglo `registros`. La forma exacta de los campos depende de `tipo`; un registro con datos incompletos o inválidos se responde `rechazado` sin frenar el resto del lote — nunca un 422 para el lote completo. `cierre_trabajo`/`cierre_sesion` (HU-05) MUTAN una fila existente en vez de crear una nueva: `uuid_cliente` identifica el EVENTO de cierre — distinto del `uuid_cliente` de apertura del trabajo/sesión que referencian. `recepcion_caldo` (HU-10 redefinida por CR-01, tarea 18) registra volumen de caldo entregado por el cliente — nunca su composición: eso es lo que ahora registra `mezcla` por separado (ver abajo), no `recepcion_caldo`. `mezcla` (espec §7, HU-78, tarea 94, revierte CR-01 del 1/9/2026) registra qué productos y en qué cantidad se cargaron en el caldo al crear una aplicación — cabecera (`trabajo_uuid_cliente`, `hora`) más una lista `productos` (nombre de texto libre, cantidad, unidad); nunca dosis por hectárea, orden de incorporación ni compatibilidad entre productos (§7.1 sigue vigente en eso, ver espec §7). `incidencia` (HU-08, tarea 22) registra un evento puntual de la sesión (caldo/ESC/batería/mecánica/clima/otro) con foto SIEMPRE obligatoria, referenciada por `uuid_cliente` a una evidencia ya subida por `POST /api/evidencias` con `tipo: foto_incidencia`. `recarga` (HU-13, tarea 23) registra cada ciclo de cambio de batería/recarga de caldo durante el vuelo — temperatura de batería > 50°C se persiste con `alerta_temperatura = true` sin rechazar el registro; independiente del evento `mezcla` (sin vínculo entre ambos en esta tarea). `estadia_entrada`/`estadia_salida` (HU-51, tarea 74) registran cuándo un equipo de trabajo llega y se va de una hacienda — sin `campania_id` (la estadía es del campo, no de una campaña) y sin verificación de pertenencia a una persona (es del equipo). `estadia_salida` referencia la estadía por el `uuid_cliente` de su `estadia_entrada`, igual criterio que `cierre_trabajo`/`cierre_sesion`. `evidencia_equipo` (HU-80, tarea 86) registra el "Reporte de Equipos": horas de vuelo declaradas del dron y las tres fotos de chequeo (control, ciclo de batería y balanceo, dron limpio), cada una referenciada por `uuid_cliente` a una evidencia ya subida por `POST /api/evidencias` — las tres son obligatorias, falta cualquiera y se rechaza el registro completo.'
      required:
        - tipo
        - uuid_cliente
      properties:
        tipo:
          description: "Adaptador delgado (ADR 0008): valida el sobre del lote, invoca el caso de\nuso y devuelve la respuesta — ninguna regla de negocio vive acá."
          type: string
          example: trabajo
          enum:
            - trabajo
            - recepcion_caldo
            - evidencia_equipo
            - mezcla
            - sesion
            - condiciones
            - incidencia
            - recarga
            - cierre_trabajo
            - cierre_sesion
            - estadia_entrada
            - estadia_salida
        uuid_cliente:
          type: string
          example: a1b2c3d4-0000-4000-8000-000000000001
        orden_id:
          description: '`trabajo`: id de servidor de la orden (del pull de catálogo).'
          type: integer
          example: 1
        lote_id:
          description: '`trabajo`: id de servidor del lote (del pull de catálogo).'
          type: integer
          example: 3
        nro_aplicacion:
          description: '`trabajo`.'
          type: integer
          example: 1
        trabajo_uuid_cliente:
          description: '`sesion`/`cierre_trabajo`/`recepcion_caldo`/`evidencia_equipo`/`mezcla`: `uuid_cliente` de apertura del trabajo — nunca el id de servidor, que puede no existir todavía si el trabajo llegó en este mismo lote.'
          type: string
          example: a1b2c3d4-0000-4000-8000-000000000001
        sesion_uuid_cliente:
          description: '`cierre_sesion`/`condiciones`/`incidencia`: `uuid_cliente` de apertura de la sesión referenciada.'
          type: string
          example: a1b2c3d4-0000-4000-8000-000000000002
        secuencia:
          description: '`sesion`: orden de apertura. `recarga`: orden de la recarga dentro de la sesión.'
          type: integer
          example: 1
        piloto_id:
          description: '`sesion`: id de servidor de la persona (del pull de catálogo).'
          type: integer
          example: 5
        auxiliar_id:
          description: '`sesion`, opcional.'
          type: integer
          example: null
          nullable: true
        dron_id:
          description: '`sesion`, opcional (HU-07): id de servidor del dron (catálogo mínimo de `Operaciones`, sin pull de catálogo propio todavía).'
          type: integer
          example: null
          nullable: true
        hectareas_declaradas:
          description: 'DECIMAL como string (invariante 6). `0` si se omite.'
          type: string
          example: '0'
        hectarea_inicial_acumulada:
          description: '`sesion`, opcional (HU-07): acumulado de DJI al abrir la sesión, control de doble conteo (espec §5). DECIMAL como string.'
          type: string
          example: null
          nullable: true
        inicio:
          type: string
          format: date-time
          example: '2026-09-01T10:00:00-04:00'
        fin:
          description: 'Requerido en `cierre_trabajo`/`cierre_sesion`.'
          type: string
          format: date-time
          example: null
          nullable: true
        litros:
          description: '`recepcion_caldo`: litros entregados por el cliente. DECIMAL como string (invariante 6).'
          type: string
          example: null
          nullable: true
        entregado_por:
          description: '`recepcion_caldo`: quién, del lado del cliente, entregó el caldo (espec §7.2).'
          type: string
          example: null
          nullable: true
        hora:
          description: '`recepcion_caldo`: cuándo se entregó. `mezcla`: cuándo se cargó el caldo. `incidencia`: cuándo ocurrió. `recarga`: cuándo ocurrió el ciclo de cambio de batería/recarga de caldo.'
          type: string
          format: date-time
          example: null
          nullable: true
        productos:
          description: '`mezcla`: lista de productos cargados en el caldo. Cada elemento trae `producto` (nombre de texto libre, transcripto del envase — sin catálogo previo que la app baje), `cantidad` (DECIMAL positivo como string, invariante 6) y `unidad` (`l`, `ml`, `kg` o `g`). No puede venir vacía: si algún elemento es inválido, se rechaza el registro `mezcla` completo.'
          type: array
          items:
            required:
              - producto
              - cantidad
              - unidad
            properties:
              producto:
                type: string
                example: 'Glifosato 48%'
              cantidad:
                type: string
                example: '2.50'
              unidad:
                type: string
                example: l
                enum:
                  - l
                  - ml
                  - kg
                  - g
            type: object
          nullable: true
        litros_consumidos:
          description: '`cierre_sesion`, opcional: litros de caldo efectivamente rociados en la sesión (espec §7.2). DECIMAL como string.'
          type: string
          example: null
          nullable: true
        litros_sobrante:
          description: '`cierre_trabajo`, opcional: litros que quedaron sin aplicar al cerrar el trabajo (espec §7.2). DECIMAL como string.'
          type: string
          example: null
          nullable: true
        evidencia_imagen_campo_uuid_cliente:
          description: '`cierre_trabajo`: REQUERIDO — `uuid_cliente` de una evidencia ya subida por `POST /api/evidencias` con `tipo: imagen_campo` (espec §9/§10, HU-09). Sin ella, referenciando una evidencia inexistente o de otro tipo, el cierre se rechaza ("sin captura no cierra").'
          type: string
          example: null
          nullable: true
        motivo_cierre:
          description: '`cierre_sesion`: catálogo espec §4.3.'
          type: string
          example: null
          nullable: true
          enum:
            - completado
            - relevo_piloto
            - cambio_dron
            - falla_equipo
            - clima
            - fin_jornada
            - otro
        momento:
          description: '`condiciones`: solo `inicio_sesion` — condiciones climáticas capturadas EN el momento de una incidencia es un concepto distinto del registro `incidencia` en sí (que vive en su propia tabla), fuera de alcance de este catálogo.'
          type: string
          example: null
          nullable: true
          enum:
            - inicio_sesion
        viento_kmh:
          description: '`condiciones`.'
          type: string
          example: '12.50'
        temperatura_c:
          description: '`condiciones`.'
          type: string
          example: '24.00'
        humedad_pct:
          description: '`condiciones`.'
          type: string
          example: '65.00'
        observacion_agronomo:
          description: '`condiciones`: obligatoria junto con `firma_observacion` cuando alguna medición cae fuera de rango (viento > 17 km/h, temperatura > 30°C o humedad > 90%) — sin ambas, el registro se rechaza.'
          type: string
          example: null
          nullable: true
        firma_observacion:
          description: '`condiciones`: sin evidencia real todavía (TE-07), texto plano.'
          type: string
          example: null
          nullable: true
        tipo_incidencia:
          description: '`incidencia`: catálogo espec §4.3. Campo separado de `tipo` (que en este objeto es el tipo de REGISTRO del lote) para evitar que ambos colisionen en la misma key.'
          type: string
          example: null
          nullable: true
          enum:
            - caldo
            - esc
            - bateria
            - mecanica
            - clima
            - otro
        descripcion:
          description: '`incidencia`, opcional: texto libre.'
          type: string
          example: null
          nullable: true
        evidencia_foto_uuid_cliente:
          description: '`incidencia`: REQUERIDO — `uuid_cliente` de una evidencia ya subida por `POST /api/evidencias` con `tipo: foto_incidencia`. Sin ella, referenciando una evidencia inexistente, de otro tipo, o ya usada por otra incidencia, el registro se rechaza (HU-08: "con foto").'
          type: string
          example: null
          nullable: true
        litros_caldo:
          description: '`recarga`: litros de caldo cargados en ese ciclo. DECIMAL como string (invariante 6).'
          type: string
          example: null
          nullable: true
        bateria_saliente_id:
          description: '`recarga`: identificador de la batería que se retira (texto libre, sin catálogo de baterías en el esquema).'
          type: string
          example: null
          nullable: true
        temperatura_bateria_c:
          description: '`recarga`: temperatura medida de la batería saliente. > 50°C persiste con `alerta_temperatura = true`, sin rechazar el registro.'
          type: string
          example: null
          nullable: true
        motivo_retraso_caldo:
          description: '`recarga`, opcional: motivo del retraso o rechazo por calidad del caldo — solo se completa si hubo retraso, no en cada recarga.'
          type: string
          example: null
          nullable: true
          enum:
            - filtro_tapado
            - grumos
            - decantacion
            - espuma
            - color_olor_anormal
        hora_retraso:
          description: '`recarga`, opcional: cuándo ocurrió el retraso/rechazo por caldo.'
          type: string
          format: date-time
          example: null
          nullable: true
        litros_combustible_generador:
          description: '`recarga`, opcional: litros de combustible cargados al generador en ese ciclo. Informativo, sin costeo (Fase 3). DECIMAL como string.'
          type: string
          example: null
          nullable: true
        equipo_trabajo_id:
          description: '`estadia_entrada`: id de servidor del equipo de trabajo (del pull de catálogo).'
          type: integer
          example: null
          nullable: true
        campo_id:
          description: '`estadia_entrada`: id de servidor del campo (del pull de catálogo).'
          type: integer
          example: null
          nullable: true
        entrada:
          description: '`estadia_entrada`: cuándo llegó el equipo.'
          type: string
          format: date-time
          example: null
          nullable: true
        vehiculo_id:
          description: '`estadia_entrada`, opcional: id de servidor del vehículo declarado.'
          type: integer
          example: null
          nullable: true
        observacion:
          description: '`estadia_entrada`, opcional: texto libre.'
          type: string
          example: null
          nullable: true
        estadia_uuid_cliente:
          description: '`estadia_salida`: `uuid_cliente` de apertura de la estadía (el `estadia_entrada` que se cierra) — nunca el id de servidor, que puede no existir todavía si la entrada llegó en este mismo lote.'
          type: string
          example: null
          nullable: true
        salida:
          description: '`estadia_salida`: cuándo se fue el equipo. Anterior o igual a la `entrada` de la estadía rechaza el registro.'
          type: string
          format: date-time
          example: null
          nullable: true
        horas_vuelo_dron:
          description: '`evidencia_equipo`: horas de vuelo del dron declaradas por el auxiliar/piloto (no calculadas). DECIMAL como string (invariante 6).'
          type: string
          example: null
          nullable: true
        foto_control_uuid_cliente:
          description: '`evidencia_equipo`, obligatorio: `uuid_cliente` de una evidencia ya subida por `POST /api/evidencias` con `tipo: foto_control`.'
          type: string
          example: null
          nullable: true
        foto_ciclo_bateria_balanceo_uuid_cliente:
          description: '`evidencia_equipo`, obligatorio: `uuid_cliente` de una evidencia con `tipo: foto_ciclo_bateria_balanceo`.'
          type: string
          example: null
          nullable: true
        foto_dron_limpio_uuid_cliente:
          description: '`evidencia_equipo`, obligatorio: `uuid_cliente` de una evidencia con `tipo: foto_dron_limpio`.'
          type: string
          example: null
          nullable: true
      type: object
    ResultadoSync:
      title: 'Resultado de aplicar un registro del lote'
      description: 'Uno por registro de entrada, en el mismo orden (espec §2.1, punto 3).'
      required:
        - uuid_cliente
        - tipo
        - estado
      properties:
        uuid_cliente:
          type: string
          example: a1b2c3d4-0000-4000-8000-000000000001
          nullable: true
        tipo:
          type: string
          example: trabajo
          nullable: true
        estado:
          type: string
          example: aplicado
          enum:
            - aplicado
            - duplicado
            - rechazado
        motivo:
          description: 'Presente solo cuando `estado` es `rechazado`.'
          type: string
          example: 'el trabajo referenciado no existe todavía'
      type: object
  securitySchemes:
    tokenDispositivo:
      type: http
      description: 'Token por dispositivo emitido por `POST /api/auth/token` (HU-03). Se envía como `Authorization: Bearer {token}`. No caduca por tiempo: vale hasta que se lo revoque desde el panel, o hasta que la cuenta o el rol con el que se emitió dejen de estar vigentes.'
      scheme: bearer
...